Use `getOrganisationsFromProfessions`
We now use this function where we can. In testing, it seems simple enough to not be worth providing a mock implementation, but we want to ensure it is called, so we use `jest.spyOn`. When using `spyOn` like this, we need to call `restoreAllMocks` in our `afterEach` to ensure tests are not influenced by previous tests