Add test for viewing only mandatory fields
The flow shouldn't break if a user has entered all mandatory fields, but
no optional ones. This is just a safety net for that on a Profession.

To prevent adding a new profession breaking the e2e tests in future,
I've added a count of professions with the desired statuses to a number
of other specs here so we don't have to increase the hard-coded value
every time.