Set slug when publishing an Org from CYA page
This was using the original "edit" journey, which copied the slug across
from an existing Organisation, but if we're publishing a brand new
organisation, there's no slug to copy, so we need to set the slug
manually.

This adds a test to cover this behaviour, and checks users can access
brand newly-published Organisations that have been published this way
(before this change, they'd be taken to an error page as there was no
slug in the url).