Set slug when publishing Profession from CYA page
This was using the original "edit" journey, which copied the slug across
from an existing Profession, but if we're publishing a brand new
Profession, there's no slug to copy, so we need to set the slug
manually.

This adds a test to cover this behaviour, and checks users can access
brand newly-published Professions that have been published this way
(before this change, they'd be taken to an error page as there was no
slug in the url).