Hide name field when submitting org with errors
Before this change, a slug value wouldn't be sent on a form submittion
error due to the way we use a `ValidationExceptionFilter` here (rather
than passing the values through like we do on the Profession side of
things).

On an error, the page is reloaded, but as the `slug` value is null,
`slug|length == 0` always evaluates to `true` and the `name` field is
shown.

This approach is a bit of a quick, hacky fix - as I think ideally we'd
either look up the Organisation from the database and handle this like
we do for Professions (which is a big change and feels a bit risky to do
at this stage), or modify the `ValidationExceptionFilter` to look up the
entity for us (which I haven't been able to do yet).

I've had to add a `submittedName` field here to also correctly render
the `name` of the organisation, otherwise that comes through as blank.