Use javascript to populate the backLink if a hash
If the backLink is a hash (i.e. if it matches the
request URL or the referrer), then we hide it
initially, and use Javascript to trigger a call
to `history.back()`. This is not an ideal
situation, but is progressively enhanced and
won’t completely break the journey for users
without Javascript.