Add "clear filters" button when we show filters
This reloads the page with no query params, so all filters are
unselected.

I've moved the entity-specific translation text for the "Filter results"
button into the top-level `app.json` here to nest it alongside the new
"clear filters" button, as we're showing the same text regardless of the
entity being searched.