Hide filters on mobile once users have searched
This prevents them having to scroll through the filters they've just
selected so they can get to their results straight away.

This isn't the most responsive solution, as if users change the size of
their screen from mobile to desktop (which feels like an unlikely
scenario), they'll need to refresh the page as it doesn't re-render with
an event listener or anything like that, but I think it's suitable for
now.