List users alphabetically on admin index page
This was more involved than expected, as we need to account for lower
case names e.g. the `beis-rpr` user, which we can't do with TypeORM's
repository `find` method. We've had to write our own query here to allow
us to `LOWER(name)` which is a bit of a pain (particularly with tests)
but does what we want.