Add other countries recognition to Qualifications
We removed a couple of these fields fairly recently, as we weren't sure
we had the best approach for capturing this information.

We're re-adding `otherCountriesRecognition[Summary]` and
`otherCountriesRecognitionUrl` alongside the new enum field
`otherCountriesRecognitionRoutes` which should help collect the data we
need until we develop a more complex iteration of this.