Present telephone numbers on external pages as international numbers
We present numbers in +44 (0)XXX format as recommended by gov.uk guidelines. As a final check, we use `libphonenumber` to check that our output telephone number represents the same number as our input telephone number

We add a `formatTelephone` function, and also expose this as a Nunjucks filter. We do not format numbers as a "tel:" link, as guidelines state to err on the side of caution and not use these when we don't know that the user's device is capable to dialling telephone numbers. This may be something we want to add later using user agent detection