Output html from intro translation
We were previously trying to translate an array in the template here,
which led to strangely outputed characters when we updated
`nestjs-i18n` here.

We may want to convert this template to not use translated text at all,
but for now, this allows us to update the version of `nestjs-i18n`.