Don't break admin when no professions to org found
There's currently an issue with the underlying data when setting a new
Regulator on a Profession, where the ProfessionToOrganisation isn't
removed from the Regulator, but its `Profession` is set to `null`. This
was causing the view to break.

As a temporary fix to this issue, while we fix the underlying data,
we're allowing the page to render without errors.