Delete orphaned rows when updating relations
We’ve had an issue where when changing an
profession’s organisations, where, rather than deleting
the whole relation, the profession was getting set to
`null` and the relation was staying around. The resulted
in some strange behaviour and some views not loading.

This added the TypeORM option `orphanedRowAction`,
which deletes any rows that are left orphaned when
saved.

I considered writing a feature spec for this, but given
#568 guards against the specific instance of the issue
that we were seeing, I didn’t want to remove that fix.