Add clientside javascript to enable adding up to 25 regulators
This change also fixes a bug where the 'Add a regulator' button still showed even when the maximum number of regulators had been added.

If the user has javascript disabled then they will be presented with a list of 25 Select boxes they can enter details of regulators into, this was judged enough of an edge case to not be an issue.

I opted to remove the elements from the DOM as opposed to just hiding them as I believe this will make the page more accessible now there are many inputs.