Add regulators to central admin decisions index
We only want to show this to central admins, as regulator users will
only see data relevant to their organisation anyway.

This follows the example for distinguishing between the overview and
single-admin views set in `ProfessionsPresenter`.

I've stopped us storing expected test results in an `expected` variable
as I find it a bit hard to read having to scroll up from the
expectation - I think we've written things this way as it gives us a bit
more of an assertion that we're testing a specific type, but I don't
think that's necessary here.