Allow users to publish decisions without editing
This adds a button to the "show" page that takes the user to
confirmation page where they see the details of the data they were
about to publish, before they can confirm this.

We may want to add this "confirmation" step to the "edit" journey too,
but this is the first step.