Allow org users to submit decision data
There's still some work to do around the order of statuses here, i.e.
whether you should be able to publish a draft status as a central admin,
whether you should be able to submit or edit an already-published
status, but this is something we'll iterate on.

This adds a new status "submitted" to the `formatStatusHelper` and
updates the `all` and `allForOrganisation` functions in the
`decisionDatasetsService` to return submitted decision datasets too.

It also removes unnecessary `await`s from non-async checkbox presenter
specs.