Hide submit button on show page if submitted
Users should only be able to submit data if it's not been submitted yet.
However, I don't want to enforce this at controller/service level, as
once we add the "submit" button to the "edit" page, we want them to be
able to edit and submit without having to save as draft first.