Add confirmation when publishing from edit
This refactors our edit controller code to always save a draft, then
redirects to the confirmation screen. This simplifies our code somewhat
and allows us to re-use the `publication.controller` endpoint which
means we don't have to duplicate code as much.

I haven't added a `go back` button to the confirmation page here (i.e.
if you decide you don't want to publish), as there's a back link that
will take you back to the edit page, but it should be easy to add that
in future if we need one.