Validate no empty routes are submitted
We can't use the standard validation mechanism here where we use
the`@Validate` decorator in the DTO itself due to the shape of the data
structure with nested arrays, which make it near-impossible to select
the invalid field and highlight it in the table. This adds a custom
`DecisionDataValidator` class that wrangles things into the right shape
so we can use the existing `ValidationFailedError` class. We highlight
the missing fields in the table.

For the tests here, I've opted not to mock the error classes for a few
reasons:

- We're using other controller methods to test validator logic (even if
  that's the only way possible using decorator methods)
- I don't want to have loads of slow end to end tests to test this logic
- In an alternative universe, our controller tests would be more
  integration-style request specs ([see what these look like in
  Rails](https://relishapp.com/rspec/rspec-rails/docs/request-specs/request-spec)
  that actually test integrations between the different classes here.
  Nest calls these "end to end" tests. We've taken to a bit of a hybrid
  approach here as we've gone with the Nest JS documentation which
  suggests controllers should be tested as individual units, but I think
  integration tests would be safer overall.