Add backlink from decision data edit page
I have had to add a 'fromEdit' query param to the url when editing data
as it is also possible to reach the page when adding new decisions data.
To test that when a user is adding decision data (not editing) the back
button takes them to the correct place I have had to assert that the
year of the decision data isn't in the url as this is the simplest
difference between the the urls.