Update seeds
We need two profession datasets for the org editor to be able to filter down to one. As the org editor account uses the Department for Education and in order not to interfere with the existing tests for DfE I have added a new profession to DfE (Primary school teacher) and a new datasets to that profession.