Revert sass-loader to 12.6.0
sass-loader 13.0.0 is incompatible with @symfony/webpack-encore 2.1.0 and shows errors in the console when building assets.
I assumed, incorrectly, that the errors would be fixed when we upgraded @symfony/webpack-encore from 1.8.2 to 2.1.0.
Although this doesn't *seem* to cause issues at the moment it may do in the future therefore I think it is safer to revert